Move worker stuff into compiler subfolder
