+++ npm/src/compiler/compiler.ts
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type { IDiagnostic, ICompletionList } from "../lib/node/qsc_wasm.cjs";
-import { log } from "./log.js";
-import { eventStringToMsg, mapDiagnostics, VSDiagnostic } from "./common.js";
-import { IQscEventTarget, QscEvents, makeEvent } from "./events.js";
+import type { IDiagnostic, ICompletionList } from "../../lib/node/qsc_wasm.cjs";
+import { log } from "../log.js";
+import { eventStringToMsg, mapDiagnostics, VSDiagnostic } from "../common.js";
+import { IQscEventTarget, QscEvents, makeEvent } from "../events.js";
 
 // The wasm types generated for the node.js bundle are just the exported APIs,
 // so use those as the set used by the shared compiler
-type Wasm = typeof import("../lib/node/qsc_wasm.cjs");
+type Wasm = typeof import("../../lib/node/qsc_wasm.cjs");
 
 // These need to be async/promise results for when communicating across a WebWorker, however
 // for running the compiler in the same thread the result will be synchronous (a resolved promise).
