+++ npm/src/compiler/worker-common.ts
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { log } from "./log.js";
-import { ICompletionList } from "../lib/web/qsc_wasm.js";
-import { DumpMsg, MessageMsg, VSDiagnostic } from "./common.js";
+import { log } from "../log.js";
+import { ICompletionList } from "../../lib/web/qsc_wasm.js";
+import { DumpMsg, MessageMsg, VSDiagnostic } from "../common.js";
 import { CompilerState, ICompiler, ICompilerWorker } from "./compiler.js";
-import { CancellationToken } from "./cancellation.js";
-import { IQscEventTarget, QscEventTarget, makeEvent } from "./events.js";
+import { CancellationToken } from "../cancellation.js";
+import { IQscEventTarget, QscEventTarget, makeEvent } from "../events.js";
 
 /*
 The WorkerProxy works by queuing up requests to send over to the Worker, only
